Added Project Support.
You have to put the path to your project file. It's not the best solution but I was too lazy to go snatch my DS Creator project handler.
I'll probably go grab it today at some point.

Note:
I am programming it on RPi4. DS Creator is on my Windows PC. So I am just too lazy to move it from one to the other. I'll try and move if with a flash drive today.